+++ tests/Alchemy/Tests/Phrasea/Core/Provider/LocaleServiceProviderTest.php
@@ -18,6 +18,55 @@ class LocaleServiceProvidertest extends \PhraseanetPHPUnitAbstract
         $this->assertEquals(Application::getAvailableLanguages(), $app['locales.available']);
     }
 
+    public function testLocalesAvailableCustomized()
+    {
+        $app = new Application();
+        $app->register(new LocaleServiceProvider());
+        $app['phraseanet.configuration'] = $this->getMock('Alchemy\Phrasea\Core\Configuration\ConfigurationInterface');
+        $app['phraseanet.configuration']->expects($this->any())
+            ->method('offsetExist')
+            ->with('main')
+            ->will($this->returnValue(true));
+        $app['phraseanet.configuration']->expects($this->any())
+            ->method('isSetup')
+            ->will($this->returnValue(true));
+        $app['phraseanet.configuration']->expects($this->any())
+            ->method('offsetGet')
+            ->with('main')
+            ->will($this->returnValue(array('languages' => array('fr_FR', 'en_US', 'de'))));
+
+        $original = Application::getAvailableLanguages();
+        unset($original['en_GB']);
+        unset($original['nl_NL']);
+
+        $this->assertEquals($original, $app['locales.available']);
+    }
+
+    public function testLocalesCustomizedWithError()
+    {
+        $app = new Application();
+        $app->register(new LocaleServiceProvider());
+        $app['phraseanet.configuration'] = $this->getMock('Alchemy\Phrasea\Core\Configuration\ConfigurationInterface');
+        $app['phraseanet.configuration']->expects($this->any())
+            ->method('offsetExist')
+            ->with('main')
+            ->will($this->returnValue(true));
+        $app['phraseanet.configuration']->expects($this->any())
+            ->method('isSetup')
+            ->will($this->returnValue(true));
+        $app['phraseanet.configuration']->expects($this->any())
+            ->method('offsetGet')
+            ->with('main')
+            ->will($this->returnValue(array('languages' => array('en_US'))));
+        $app['monolog'] = $this->getMock('Psr\Log\LoggerInterface');
+        $app['monolog']->expects($this->once())
+            ->method('error');
+
+        $original = Application::getAvailableLanguages();
+
+        $this->assertEquals($original, $app['locales.available']);
+    }
+
     public function testLocalesI18nAvailable()
     {
         $app = new Application();